FEATURE :sparkles: Only allow author (and admins) to edit own articles

+++ app/Http/Controllers/Admin/ArticleController.php
@@ -10,7 +10,11 @@ class ArticleController extends Controller
 {
     public function index()
     {
-        $articles = Article::all();
+        if(auth()->user()->is_admin) {
+            $articles = Article::all();
+        } else {
+            $articles = Article::where('author_id', auth()->id())->get();
+        }
 
         return view('admin.articles.index', ['articles' => $articles]);
     }
@@ -47,11 +51,19 @@ class ArticleController extends Controller
 
     public function edit(Article $article)
     {
+        if(!$article->canEdit()) {
+            abort(401);
+        }
+
         return view('admin.articles.edit', compact('article'));
     }
 
     public function update(Request $request, Article $article)
     {
+        if(!$article->canEdit()) {
+            abort(401);
+        }
+
         $request->validate([
             'title' => ['required', 'string', 'min:10', 'max:255'],
             'content' => ['required'],
@@ -71,6 +83,10 @@ class ArticleController extends Controller
 
     public function destroy(Request $request, Article $article)
     {
+        if(!$article->canEdit()) {
+            abort(401);
+        }
+
         $article->delete();
 
         return redirect()->route('admin.articles.index');
